+++ src/custom/state/orders/updaters/utils.ts
@@ -1,0 +1,78 @@
+import { Order, OrderFulfillmentData } from 'state/orders/actions'
+import { getOrder, OrderID, OrderMetaData } from 'utils/operator'
+import { stringToCurrency } from 'state/swap/extension'
+import { formatSmart } from 'utils/format'
+import { SHORT_PRECISION } from 'constants/index'
+import { ApiOrderStatus, classifyOrder } from 'state/orders/utils'
+import { SupportedChainId as ChainId } from 'constants/chains'
+
+export type OrderLogPopupMixData = OrderFulfillmentData | OrderID
+
+function _computeFulfilledSummary({
+  orderFromStore,
+  orderFromApi,
+}: {
+  orderFromStore?: Order
+  orderFromApi: OrderMetaData | null
+}) {
+  // Default to store's current order summary
+  let summary: string | undefined = orderFromStore?.summary
+
+  // if we can find the order from the API
+  // and our specific order exists in our state, let's use that
+  if (orderFromApi) {
+    const { buyToken, sellToken, executedBuyAmount, executedSellAmount } = orderFromApi
+
+    if (orderFromStore) {
+      const { inputToken, outputToken } = orderFromStore
+      // don't show amounts in atoms
+      const inputAmount = stringToCurrency(executedSellAmount, inputToken)
+      const outputAmount = stringToCurrency(executedBuyAmount, outputToken)
+
+      summary = `Swap ${formatSmart(inputAmount, SHORT_PRECISION)} ${inputAmount.currency.symbol} for ${formatSmart(
+        outputAmount,
+        SHORT_PRECISION
+      )} ${outputAmount.currency.symbol}`
+    } else {
+      // We only have the API order info, let's at least use that
+      summary = `Swap ${sellToken} for ${buyToken}`
+    }
+  } else {
+    console.log(`[state:orders:updater] computeFulfilledSummary::API data not yet in sync with blockchain`)
+  }
+
+  return summary
+}
+
+type PopupData = {
+  status: ApiOrderStatus
+  popupData?: OrderLogPopupMixData
+}
+
+export async function fetchOrderPopupData(orderFromStore: Order, chainId: ChainId): Promise<PopupData> {
+  const orderFromApi = await getOrder(chainId, orderFromStore.id)
+
+  const status = classifyOrder(orderFromApi)
+
+  let popupData = undefined
+
+  switch (status) {
+    case 'fulfilled':
+      popupData = {
+        id: orderFromStore.id,
+        fulfillmentTime: new Date().toISOString(),
+        transactionHash: '', // there's no need  for a txHash as we'll link the notification to the Explorer
+        summary: _computeFulfilledSummary({ orderFromStore, orderFromApi }),
+      }
+      break
+    case 'expired':
+    case 'cancelled':
+      popupData = orderFromStore.id
+      break
+    default:
+      // No popup for other states
+      break
+  }
+
+  return { status, popupData }
+}
