+++ src/index.tsx
@@ -20,7 +20,7 @@ import TransactionUpdater from './state/transactions/updater'
 import UserUpdater from './state/user/updater'
 import FeesUpdater from 'state/price/updater'
 import GasUpdater from 'state/gas/updater'
-import { CancelledOrdersUpdater, EventUpdater, UnfillableOrdersUpdater } from 'state/orders/updater'
+import { CancelledOrdersUpdater, EventUpdater, UnfillableOrdersUpdater } from 'state/orders/updaters'
 // import { EventUpdater } from 'state/orders/mocks'
 import ThemeProvider, { FixedGlobalStyle, ThemedGlobalStyle } from 'theme'
 import getLibrary from './utils/getLibrary'
